# zstd: Speed up when WithDecoderLowmem(false)